type-c-interface: Move controller trait into own module, clean-up types
Move the `Controller` trait into its own module, and create a dedicated
module for types used as arguments/return values for various PD
functions.
