+++ examples/std/src/bin/type_c/ucsi.rs
@@ -22,8 +22,9 @@ use power_policy_service::psu::PsuEventReceivers;
 use power_policy_service::service::registration::ArrayRegistration;
 use static_cell::StaticCell;
 use std_examples::type_c::mock_controller::{self, InterruptReceiver, Port};
+use type_c_interface::controller::ControllerId;
 use type_c_interface::port::event::PortEventBitfield;
-use type_c_interface::port::{ControllerId, Device, PortRegistration};
+use type_c_interface::port::{Device, PortRegistration};
 use type_c_interface::service::context::Context;
 use type_c_interface::service::event::{PortEvent as ServicePortEvent, PortEventData as ServicePortEventData};
 use type_c_service::bridge::Bridge;
