+++ type-c-interface/src/control/mod.rs
@@ -1,0 +1,9 @@
+//! Shared types for controlling a PD port
+pub mod dp;
+pub mod pd;
+pub mod power;
+pub mod retimer;
+pub mod tbt;
+pub mod type_c;
+pub mod usb;
+pub mod vdm;
