+++ type-c-interface/src/control/retimer.rs
@@ -1,0 +1,11 @@
+//! Retimer related control types
+
+/// Retimer update state
+#[derive(Copy, Clone, Debug, PartialEq)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub enum RetimerFwUpdateState {
+    /// Retimer FW Update Inactive
+    Inactive,
+    /// Retimer FW Update Active
+    Active,
+}
