+++ type-c-interface/src/control/tbt.rs
@@ -1,0 +1,9 @@
+//! Thunderbolt-related control types
+
+/// Thunderbolt control configuration
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+#[derive(Debug, Clone, Default, Copy, PartialEq)]
+pub struct TbtConfig {
+    /// Enable Thunderbolt
+    pub tbt_enabled: bool,
+}
