+++ type-c-interface/src/lib.rs
@@ -1,4 +1,6 @@
 //! Interface for type-C service.
 #![no_std]
+pub mod control;
+pub mod controller;
 pub mod port;
 pub mod service;
